requirements.txt
django-htmx==1.6.0
django-modelcluster==5.2
django-recaptcha==2.0.6
django-taggit==1.5.1
django-tailwind==3.0.1
emoji==0.6.0
